Bevel effect finished, memory issues fixed, HSB sliders added.

+++ AmazeKit/AmazeKit/UIImage+AKPixelData.h
@@ -12,10 +12,10 @@
 
 // Helper C functions
 typedef struct {
-	CGFloat red;
-	CGFloat blue;
-	CGFloat green;
-	CGFloat alpha;
+	uint8_t red;
+	uint8_t blue;
+	uint8_t green;
+	uint8_t alpha;
 } AKPixelData;
